requirements.txt
hydralit-components==1.0.10
Markdown==3.6
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==9.3.0
protobuf==3.20.1
stqdm==0.0.5
streamlit==1.32.2
streamlit-lottie==0.0.5
tqdm==4.66.2
matplotlib==3.7.2
datasets
transformers
joblib
mysql-connector-python